+++ src/types.ts
@@ -1,4 +1,84 @@
-import { JWK, JWTHeaderParameters, JWTPayload } from 'jose';
+export interface JWK {
+  /** JWK "alg" (Algorithm) Parameter. */
+  alg?: string;
+  crv?: string;
+  d?: string;
+  dp?: string;
+  dq?: string;
+  e?: string;
+  /** JWK "ext" (Extractable) Parameter. */
+  ext?: boolean;
+  k?: string;
+  /** JWK "key_ops" (Key Operations) Parameter. */
+  key_ops?: string[];
+  /** JWK "kid" (Key ID) Parameter. */
+  kid?: string;
+  /** JWK "kty" (Key Type) Parameter. */
+  kty?: string;
+  n?: string;
+  oth?: Array<{
+    d?: string;
+    r?: string;
+    t?: string;
+  }>;
+  p?: string;
+  q?: string;
+  qi?: string;
+  /** JWK "use" (Public Key Use) Parameter. */
+  use?: string;
+  x?: string;
+  y?: string;
+  /** JWK "x5c" (X.509 Certificate Chain) Parameter. */
+  x5c?: string[];
+  /** JWK "x5t" (X.509 Certificate SHA-1 Thumbprint) Parameter. */
+  x5t?: string;
+  /** "x5t#S256" (X.509 Certificate SHA-256 Thumbprint) Parameter. */
+  'x5t#S256'?: string;
+  /** JWK "x5u" (X.509 URL) Parameter. */
+  x5u?: string;
+  [propName: string]: unknown;
+}
+
+export interface JWTHeaderParameters {
+  /** "kid" (Key ID) Header Parameter. */
+  kid?: string;
+  /** "x5t" (X.509 Certificate SHA-1 Thumbprint) Header Parameter. */
+  x5t?: string;
+  /** "x5c" (X.509 Certificate Chain) Header Parameter. */
+  x5c?: string[];
+  /** "x5u" (X.509 URL) Header Parameter. */
+  x5u?: string;
+  /** "jku" (JWK Set URL) Header Parameter. */
+  jku?: string;
+  /** "jwk" (JSON Web Key) Header Parameter. */
+  jwk?: Pick<JWK, 'kty' | 'crv' | 'x' | 'y' | 'e' | 'n'>;
+  /** "typ" (Type) Header Parameter. */
+  typ?: string;
+  /** "cty" (Content Type) Header Parameter. */
+  cty?: string;
+  /** JWS "crit" (Critical) Header Parameter. */
+  crit?: string[];
+  /** Any other JWS Header member. */
+  [propName: string]: unknown;
+  /** JWS "alg" (Algorithm) Header Parameter. */
+  alg: string;
+  /**
+   * This JWS Extension Header Parameter modifies the JWS Payload representation and the JWS Signing
+   * Input computation as per {@link https://www.rfc-editor.org/rfc/rfc7797 RFC7797}.
+   */
+  b64?: true;
+}
+
+export interface JWTPayload {
+  iss?: string;
+  sub?: string;
+  aud?: string | string[];
+  jti?: string;
+  nbf?: number;
+  exp?: number;
+  iat?: number;
+  [propName: string]: unknown;
+}
 
 export interface UnverifiedJWT {
   header: JWTHeaderParameters;
@@ -69,7 +149,6 @@ export interface IssueSDJWTOptions {
 export interface VerifySdJwtOptions {
   kb?: {
     verifier?: KeyBindingVerifier;
-    skipCheck?: boolean;
   };
 }
 /**